logic done

+++ api/db/model/ledger.go
@@ -5,8 +5,8 @@ import "gorm.io/gorm"
 type LedgerEntry struct {
 	gorm.Model
 
-	AccountID uint  `gorm:"not null"`
-	Amount    int64 `gorm:"not null"`
+	AccountID uint    `gorm:"not null"`
+	Amount    float64 `gorm:"not null"`
 
 	TransactionID uint
 	TransferID    *uint
