+++ api/db/model/transaction.go
@@ -4,7 +4,7 @@ import "gorm.io/gorm"
 
 type Transaction struct {
 	gorm.Model
-	Type      string `gorm:"not null"`
-	Amount    int64  `gorm:"not null"`
-	AccountID uint   `gorm:"not null"`
+	Type      string  `gorm:"not null"`
+	Amount    float64 `gorm:"not null"`
+	AccountID uint    `gorm:"not null"`
 }
