+++ api/db/model/transfer.go
@@ -4,7 +4,7 @@ import "gorm.io/gorm"
 
 type Transfer struct {
 	gorm.Model
-	FromAccountID uint  `gorm:"not null"`
-	ToAccountID   uint  `gorm:"not null"`
-	Amount        int64 `gorm:"not null"`
+	FromAccountID uint    `gorm:"not null"`
+	ToAccountID   uint    `gorm:"not null"`
+	Amount        float64 `gorm:"not null"`
 }
